backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
jinja2==3.1.4
pydantic==2.8.2
pydantic-settings==2.4.0
SQLAlchemy==2.0.32
alembic==1.13.2
psycopg2-binary==2.9.9
python-multipart==0.0.9
httpx==0.27.0
python-jose[cryptography]==3.3.0
pdfminer.six==20240706
pytesseract==0.3.13
Pillow==10.4.0